+++ Race_Element/MainWindow.xaml.cs
@@ -212,7 +212,7 @@ public partial class MainWindow : Window
 
     private void MainWindow_Loaded(object sender, RoutedEventArgs e)
     {
-        App.Instance._startScreenOverlay.Stop(false);
+        App.Instance._startScreenOverlay.Stop();
         App.Instance._startScreenOverlay.Dispose();
 
         ThreadPool.QueueUserWorkItem(x =>
